+++ src/DependencyInjection/Configuration.php
@@ -284,8 +284,9 @@ class Configuration implements ConfigurationInterface
     private function addFormSection($nodeBuilder)
     {
         $nodeBuilder
-           ->arrayNode('form')
+            ->arrayNode('form')
                 ->addDefaultsIfNotSet()
+                ->fixXmlConfig('option')
                 ->children()
                     ->arrayNode('data_class')
                         ->addDefaultsIfNotSet()
@@ -293,8 +294,18 @@ class Configuration implements ConfigurationInterface
                             ->scalarNode('seo_metadata')->defaultNull()->end()
                         ->end()
                     ->end()
+                    ->arrayNode('options')
+                        ->addDefaultsIfNotSet()
+                        ->children()
+                            ->enumNode('generic_metadata')
+                                ->info('Whether to show fields to edit generic SEO information. Needs burgov/key-value-form-bundle.')
+                                ->values([true, false, 'auto'])
+                                ->defaultValue('auto')
+                            ->end()
+                        ->end()
+                    ->end()
                 ->end()
-           ->end()
-           ;
+            ->end()
+       ;
     }
 }